requirements.txt
flask==3.0.0
flask-socketio==5.3.6
flask-cors==4.0.0
pynput==1.8.1
python-socketio==5.11.0
pystray==0.19.5
Pillow>=12.0.0
requests>=2.31.0
packaging>=23.0
pyinstaller>=6.0.0
